extracted search list form to new component
