+++ src/views/List.jsx
@@ -1,50 +1,22 @@
 import { useState, useEffect } from 'react';
 import { ListItem } from '../components';
+import SearchList from '../components/SearchList';
+import { useParams } from 'react-router-dom';
 
 export function List({ data }) {
-	const [value, setValue] = useState('');
 	const [newList, setNewList] = useState([]);
+	const { path } = useParams();
 
 	useEffect(() => {
 		setNewList(data);
 	}, [data]);
 
-	const handlefiltering = (e) => {
-		const userinput = e.target.value.toLowerCase();
-		setValue(e.target.value);
-		let gatherItem = [];
-		for (let i = 0; i < data.length; i++) {
-			console.log('data[i].name', data[i]?.name);
-			if (data[i]?.name.toLowerCase().includes(userinput)) {
-				gatherItem.push(data[i]);
-			}
-		}
-		setNewList(gatherItem);
-	};
-
-	const resetInput = (e) => {
-		e.preventDefault();
-
-		setValue('');
-	};
-
 	return (
 		<>
-			<p>
-				Hello from the <code>/list</code> page!
-			</p>
-			<form>
-				<label htmlFor="search">Search item</label>
-				<input
-					id="search"
-					type="text"
-					onChange={(e) => handlefiltering(e)}
-					value={value}
-				/>
-				<button onClick={(e) => resetInput(e)} aria-label="clear search bar">
-					x
-				</button>
-			</form>
+			<h2>
+				You are on the <code>{path}</code> list!
+			</h2>
+			<SearchList data={data} setNewList={setNewList} />
 			<div>
 				{newList?.length === 0 ? (
 					<></>
